push all code

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,9 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            //We should user double here to get accuracy in commission_rate
+            //In float there are 82.89899999999999 lots of values after point so we need accuracy here 
+            $table->double('commission_rate');
             $table->string('discount_code');
             $table->timestamps();
         });
